consolidated JDFDuration to be based on Long rather than Integer
git-svn-id: http://svn.cip4.org/svnroot/trunk@6877 25ee8a6d-831b-0410-907a-ff8528d65d49

+++ src/test/java/org/cip4/jdflib/datatypes/JDFDurationTest.java
@@ -274,7 +274,7 @@ public class JDFDurationTest extends TestCase
 	@Test
 	public final void testGetDurationINF()
 	{
-		assertEquals(JDFDuration.createDuration("INF").getDuration(), Integer.MAX_VALUE);
+		assertEquals(JDFDuration.createDuration("INF").getDuration(), Long.MAX_VALUE);
 	}
 
 	/**
